[DataPersistence] First version of save and load framework and able to S/L event logs (#53)
To work with the save/load (data persistence) framework, you need to implement the interface IDataPersistence
Options:

if you are working with non-monoBehavior objects (example: GameEventLogger), inheritance the DataPersistence class, which has a default implementation that handles the following
registration (you need to register the class so that dataPersistenceManager knows your existence, set autoRegister to false when you want to manually manage when to register)
LoadData (default implementation that loads serialized data from saved data, deserializes, and then sets the value to all fields)
SaveData (default implementation that serializes the object and writes to save data)
GetSaveFileName (Name of object you want to save to, default is the name of class type)
Dispose (remove registration, only call it when you need to destroy your object)
Restart called when we restart the scene without loading, please implement if needed to do anything, otherwise, just leave it.
Please override the provided virtual method if it doesn't meet your needs since the default implementation is limited in functionality
If you are working with the MonoBehavior object (example PlayerManager), currently there's no dedicated base class, you need to implement IDataPersistence yourself, in the future, we can add a wrapper around monoBehavior if needed
registration you need to call the default registration implementation in the interface when the object is initialized, for example: ((IDataPersistence)this).RegisterExistence();
SaveData (you need to collect the information you need and create an object to store all the information, then call DataPersistenceManager.SaveDataFile(fileName, playerInfo); in this example, then playerInfo object of type PlayerSaveInfo). You can have each of your object implement this interface (e.g. playerManager saves position, weaponHandler saves weapon info) or have a centralized place to store all (playerManager gets info from weaponHandler and handles save/load) example is the second option
LoadData (you need to call DataPersistenceManager.LoadDataFile<PlayerSaveInfo>(fileName); to get the deserialized object, in this case, PlayerSaveInfo, then set it to the gameObject, such as set position, set animation state, etc)
GetSaveFileName I recommend just returning the name of the class type for example return nameof(PlayerSaveInfo);
OnDestory Unity method, you have to call ((IDataPersistence)this).Dispose(); to remove this class from registration, in the future, we might be able to move this to a common parent class.
Restart (as of now clear your data object to prevent loading saved data)
After loading data, since the scene will be restarted, we need to store the data to some static place, for this player example, we used Backend.Instance.PlayerStats, in future we can move to some dedicated place to store the info.
Addition things to pay attention to:

Since save and load restart the scene, thus if you have added an event handler (callbacks) to objects that don't destroy by reloading the scene (such as input map), you need to remove it on object destroy, e.g.
  private void OnDestroy()
  {
      input.Player.Look.started -= Look;
      input.Player.Look.performed -= Look;
        input.Player.Look.canceled -= Look;
    }